backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.3
drf-yasg==1.21.8
celery==5.4.0
redis==5.2.1
cryptography==45.0.0
heyoo==0.1.2
paynow==1.0.8
